+++ CHANGELOG.md
@@ -5,6 +5,20 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [0.6] - 2022-04-13
+### Added
+- `save_params_filter` function
+### Changed
+- `__del__` method updated
+- `message` field changed in `save_fig_file` function
+- `message` field changed in `save_config_file` function
+- `message` field changed in `save_data_file` function
+- `message` field changed in `nft_storage_upload` function
+- `depth` section added to config/data file
+- `linewidth` parameter added to `plot` method
+- `linewidth` parameter added to `plot_params_filter` function
+- Random mode modified
+- `README.md` updated
 ## [0.5] - 2022-03-21
 ### Added
 - `__del__` method
@@ -81,7 +95,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `generate` method
 - `nft_storage` method
 
-[Unreleased]: https://github.com/sepandhaghighi/samila/compare/v0.5...dev
+[Unreleased]: https://github.com/sepandhaghighi/samila/compare/v0.6...dev
+[0.6]: https://github.com/sepandhaghighi/samila/compare/v0.5...v0.6
 [0.5]: https://github.com/sepandhaghighi/samila/compare/v0.4...v0.5
 [0.4]: https://github.com/sepandhaghighi/samila/compare/v0.3...v0.4
 [0.3]: https://github.com/sepandhaghighi/samila/compare/v0.2...v0.3